refactor(agent): the run's own types live with AgentRun, not in rig-core
rig-core has requests, responses, messages, tools and streams — no run, no
loop, no prompt. Four things the dissolve had filed under 'vocabulary' only
mean something inside a run, so they move to rig_agent::run:

- run::response — PromptResponse, CompletionCall, PromptError (a run's result
  and its loop-level errors: MaxTurns, PromptCancelled, UnknownToolCall);
- run::policy — InvalidToolCallAction, InvalidToolCallContext, RetryRequest
  (AgentRun::resolve_invalid_tool_call's inputs; nothing in rig-core consumes
  them);
- run::streamed — the streamed-turn assembler, whose resolutions are the
  machine's and whose tests drive an AgentRun (reunited here; the copied
  helpers are gone).

What stays in rig-core is what prepare_request consumes or produces:
transcript, completion::{output, prepare, spec} and RequestPatch, now at
completion::patch. rig::run is rig_agent::run behind the agent feature; the
request vocabulary is reachable without it at rig::core::… paths. The
no-async guard follows the files; the 0.42 path pin is unchanged.
