+++ crates/rig-agent/src/agent/prompt_request/mod.rs
@@ -319,7 +319,7 @@ impl PromptRequest<Standard> {
 }
 
 /// Details for one successfully completed completion request made by an agent run.
-pub use rig_core::completion::response::{CompletionCall, PromptResponse};
+pub use crate::run::response::{CompletionCall, PromptResponse};
 pub(crate) use rig_core::transcript::{
     assistant_text_from_choice, is_empty_assistant_turn, tool_result_output,
 };
