+++ crates/rig-core/src/completion/mod.rs
@@ -30,17 +30,15 @@
 pub mod handle;
 pub mod message;
 pub mod output;
-pub mod policy;
+pub mod patch;
 pub mod prepare;
 pub mod request;
-pub mod response;
 pub mod spec;
 
 pub use handle::{ModelHandle, ModelRef};
 pub use message::{AssistantContent, Message, MessageError};
 pub use output::OutputMode;
-pub use policy::{InvalidToolCallAction, InvalidToolCallContext, RequestPatch, RetryRequest};
+pub use patch::RequestPatch;
 pub use prepare::{PrepareError, PreparedRequest, prepare_request};
 pub use request::*;
-pub use response::{CompletionCall, PromptError, PromptResponse};
 pub use spec::RunSpec;
